# backend/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Swyngora Market API
  description: |
    Public market data for Swyngora (v0.1).

    - Candles, 24h volume/ticker, and spot market listing support **binance**, **coinbase**,
      and **bybit** (select with `?exchange=`).
    - Circulating / total / max supply is sourced from the **Binance** public marketing
      symbol list (asset-level, used for all venues' mcap enrichment), refreshed daily.
      Max supply is null when undefined.
    - Watchlist is unauthenticated state keyed by client-supplied `clientId`
      (must be non-empty; not for multi-tenant production without real auth).
      Data is persisted in SQLite on the server (survives restarts). Each list has a monotonic
      `version` for multi-device sync: send `baseVersion` on writes; concurrent adds of different
      symbols auto-merge; delete-vs-update on the same symbol returns 409 with both sides.
    - Price alerts support `one_time` (default) or `repeating` modes for above/below targets.
      One-time alerts become `triggered` after the first fire; repeating alerts stay `active`
      and re-fire only after the price returns to the safe side and crosses again.
    - Clients may register a webhook URL with deliveryMode `immediate` or `hourly_digest`,
      plus optional quiet hours (IANA timeZone + local start/end). Delivery is deferred until
      quiet hours end; pending notifications survive restarts.
...
      orders, and performance), market buy/sell at last price,
...
servers:
...
paths:
...
  /api/v1/account/api-keys:
    get:
      summary: List API keys for this account
      description: |
        Named keys for bots/apps. Secrets are never listed. Manage keys with the
        main account (`X-Client-Id` plus optional process `API_AUTH_TOKEN`).
        User-issued keys cannot list or create other keys.
      operationId: listAccountAPIKeys
      tags: [Account]
      parameters:
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      responses:
        '200':
          description: Key metadata (no secrets)
          content:
            application/json:
              schema:
                type: object
                properties:
                  clientId: { type: string }
                  keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/AccountAPIKey'
                  count: { type: integer }
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
    post:
      summary: Create a named API key
      description: |
        permission=read → GET only. permission=trade → can place/cancel paper trades
        and other mutations except account close and key management.
        The full secret is returned once.
      operationId: createAccountAPIKey
      tags: [Account]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                clientId: { type: string }
                name: { type: string, example: Trading bot }
                permission: { type: string, enum: [read, trade], default: read }
      responses:
        '201':
          description: Created key including one-time secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountAPIKeyCreated'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
  /api/v1/account/api-keys/{id}:
    delete:
      summary: Revoke an API key
      description: Immediate; the secret stops working. Idempotent if already revoked.
      operationId: revokeAccountAPIKey
      tags: [Account]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      responses:
        '200':
          description: Revoked key metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountAPIKey'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
components:
  responses:
    Error:
      description: Error payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    AccountAPIKey:
      type: object
      description: Named user API key metadata (secret is never listed)
      properties:
        id: { type: string }
        name: { type: string }
        prefix: { type: string, description: Public prefix of the secret (e.g. swy_ab12cd34) }
        permission: { type: string, enum: [read, trade] }
        createdAt: { type: string, format: date-time }
        lastUsedAt: { type: string, format: date-time, nullable: true }
        revokedAt: { type: string, format: date-time, nullable: true }
        revoked: { type: boolean }
    AccountAPIKeyCreated:
      allOf:
        - $ref: '#/components/schemas/AccountAPIKey'
        - type: object
          properties:
            secret:
              type: string
              description: Full token; shown only at creation. Send as Authorization Bearer or X-API-Key.
...